queue controller and stuff pretty much done -- finish integrating
need to queue view does not have lace on window but everything seems hooked up. Add a way to be able to add an item to the queue and inside of queue have the widgets naturally being condensed but when theyre clicked on have them expand to show a more detailed thing within the same layout. When expanded have a button to edit it in detail -- low priority. Also have a button the export it individually -- higher priority.